hide cache key type from aggregators
